+++ NEWS.md
@@ -1,6 +1,7 @@
 # mlr3proba 0.1.4
 
 * Density estimation has now been added to mlr3proba, see `TaskDens`, `LearnerDens`, `PredictionDens`, and `MeasureDens`.
+* Added learners for density estimation, see `mlr3::mlr_learners$keys("^dens")` for the full list
 
 # mlr3proba 0.1.3